refactor: formatting package path
