+++ src/main/kotlin/string/calculator/dto/SplitInput.kt
@@ -1,5 +1,5 @@
 package string.calculator.dto
 
-import string.calculator.enum.Operator
+import string.calculator.domain.Operator
 
 class SplitInput(val numbers: MutableList<Long>, val operators: MutableList<Operator>)
